requirements.txt
pytesseract
pyMuPDF
opencv-python
numpy
Pillow